TSD-417: Code Improvement; fix Code Sniffer "Unescaped output detected" and "Each class must be in a file by itself"
